Move spirit standings and team page to nested navigator
