+++ src/navigation/root/tab/spirit/ScoreBoardScreen/index.ts
