+++ src/navigation/root/tab/spirit/TeamScreen/index.ts
